Declare methods 'const'
This allows you to do some casting and avoid a copy, as now shown in the
wind2rrd example.

old: n2k::WindData w(m);
new: const n2k::WindData &w = static_cast<const n2k::WindData&>(m);

This is safe because a WindData object is exactly the same size and
format as a Message. The only real difference is that there are
additional methods that can be called on the object.

This seems like a lot of work for the person implementing the handler,
so I'm going to keep looking for better ways to do this.